SonarQube batch 4: S3230 in-class init (~167), S5276 float conv (~44), S1066 collapsible if (~10), S3628 range-for (3), S1103 trailing whitespace (all files)
- S3230: Convert 15 structs in Types.h + 5 other files from constructor init-lists to in-class member initializers with = default constructors
- S5276: Add constexpr PI_F, use acosf/sinf/cosf/fmodf, explicit static_cast<float> for random()/size_t/unsigned long narrowing conversions across 8 files
- S1066: Merge collapsible if statements in ContactSensors, ChaosController, OscillationController, PursuitController
- S3628: Use range-based for loops in SequenceTableManager, FilesystemManager
- S1103: Strip trailing whitespace from all .h/.cpp files
- Build: SUCCESS, 180/180 tests PASSED

+++ src/core/filesystem/FileSystem.cpp
@@ -69,7 +69,7 @@ bool FileSystem::mount() {
     size_t usedBytes = LittleFS.usedBytes();
     String msg = "LittleFS: " + String(totalBytes / 1024) + " KB total, " +
                  String(usedBytes / 1024) + " KB used (" +
-                 String((usedBytes * 100.0f) / totalBytes, 1) + "%)";
+                 String((static_cast<float>(usedBytes) * 100.0f) / static_cast<float>(totalBytes), 1) + "%)";
     if (engine) engine->info(msg);
     else Serial.println("[FileSystem] " + msg);
   }
@@ -117,14 +117,14 @@ String FileSystem::readFileAsString(const String& path, size_t maxSize) {
 
   // Limit read size for safety
   size_t readSize = min(file.size(), maxSize);
-  
+
   // Bulk read for performance
   char* buf = new (std::nothrow) char[readSize + 1];
   if (!buf) {
     file.close();
     return "";
   }
-  
+
   size_t bytesRead = file.readBytes(buf, readSize);
   buf[bytesRead] = '\0';
   String content(buf);
@@ -189,7 +189,7 @@ float FileSystem::getDiskUsagePercent() const {
   uint32_t total = getTotalBytes();
   if (total == 0) return 0.0f;
   uint32_t used = getUsedBytes();
-  return (used * 100.0f) / total;
+  return (static_cast<float>(used) * 100.0f) / static_cast<float>(total);
 }
 
 // ============================================================================
